chore(suite): reorganize PrivacyAccount folder
