+++ packages/suite/src/components/wallet/index.tsx
@@ -5,7 +5,7 @@ import { WalletLayoutNavLink } from './WalletLayout/WalletLayoutNavLink';
 import { InputError } from './InputError';
 import { AccountExceptionLayout } from './AccountExceptionLayout';
 import { DiscoveryProgress } from './DiscoveryProgress';
-import { UtxoAnonymity } from './PrivacyAccount/UtxoAnonymity';
+import { UtxoAnonymity } from './UtxoAnonymity';
 import { Pagination } from './Pagination';
 import { TransactionTimestamp } from './TransactionTimestamp';
 import { withSelectedAccountLoaded } from './hocs';
