+++ packages/suite/src/views/wallet/details/CoinjoinLogs.tsx
