+++ packages/suite/src/views/wallet/details/CoinjoinSetup/SetupSlider/SetupSlider.tsx
