+++ packages/suite/src/views/wallet/details/CoinjoinSetup/SetupSlider/SliderInput.tsx
