+++ packages/suite/src/views/wallet/transactions/CoinjoinSummary/BalancePrivacyBreakdown/BalancePrivacyBreakdown.tsx
@@ -1,14 +1,15 @@
 import styled from 'styled-components';
+
 import { useTheme, Icon } from '@trezor/components';
 import { isZero } from '@suite-common/wallet-utils';
 import { Translation } from 'src/components/suite/Translation';
 import { useSelector } from 'src/hooks/suite';
-import { CryptoAmountWithHeader } from 'src/components/wallet/PrivacyAccount/CryptoAmountWithHeader';
 import {
     selectCurrentCoinjoinBalanceBreakdown,
     selectCurrentCoinjoinSession,
 } from 'src/reducers/wallet/coinjoinReducer';
 import { selectSelectedAccount } from 'src/reducers/wallet/selectedAccountReducer';
+import { CryptoAmountWithHeader } from './CryptoAmountWithHeader';
 
 const BalanceContainer = styled.div`
     display: flex;
